+++ Netkan/Services/IModuleService.cs
@@ -10,7 +10,7 @@ namespace CKAN.NetKAN.Services
     {
         Tuple<ZipEntry, bool> FindInternalAvc(CkanModule module, ZipFile zipfile, string internalFilePath);
         AvcVersion GetInternalAvc(CkanModule module, string filePath, string internalFilePath = null);
-        JObject GetInternalCkan(string filePath);
+        JObject GetInternalCkan(CkanModule module, string zipPath, GameInstance inst);
         bool HasInstallableFiles(CkanModule module, string filePath);
 
         IEnumerable<InstallableFile> GetConfigFiles(CkanModule module, ZipFile zip, GameInstance inst);